+++ Servers/utils/fileUpload.utils.ts
@@ -22,10 +22,11 @@ export const uploadFile = async (
     | "Management system clauses group"
     | "Reference controls group"
     | "Clauses and annexes report",
+  tenant: string,
   transaction: Transaction | null = null
 ) => {
   const projectIsDemo = await sequelize.query(
-    "SELECT is_demo FROM projects WHERE id = :id",
+    `SELECT is_demo FROM ${tenant}.projects WHERE id = :id`,
     {
       replacements: { id: project_id },
       mapToModel: true,
@@ -34,7 +35,7 @@ export const uploadFile = async (
     }
   );
   const is_demo = projectIsDemo[0].is_demo || false;
-  const query = `INSERT INTO files
+  const query = `INSERT INTO ${tenant}.files
     (
       filename, content, type, project_id, uploaded_by, uploaded_time, is_demo, source
     )
@@ -60,8 +61,8 @@ export const uploadFile = async (
   return result[0];
 };
 
-export const deleteFileById = async (id: number, transaction: Transaction) => {
-  const query = `DELETE FROM files WHERE id = :id`;
+export const deleteFileById = async (id: number, tenant: string, transaction: Transaction) => {
+  const query = `DELETE FROM ${tenant}.files WHERE id = :id`;
   const result = await sequelize.query(query, {
     replacements: { id },
     mapToModel: true,
@@ -72,8 +73,8 @@ export const deleteFileById = async (id: number, transaction: Transaction) => {
   return result.length > 0;
 };
 
-export const getFileById = async (id: number) => {
-  const query = `SELECT * FROM files WHERE id = :id`;
+export const getFileById = async (id: number, tenant: string) => {
+  const query = `SELECT * FROM ${tenant}.files WHERE id = :id`;
   const result = await sequelize.query(query, {
     replacements: { id },
     mapToModel: true,
@@ -82,7 +83,7 @@ export const getFileById = async (id: number) => {
   return result[0];
 };
 
-export const getFileMetadataByProjectId = async (project_id: number) => {
+export const getFileMetadataByProjectId = async (project_id: number, tenant: string) => {
   const query = `SELECT 
   f.id, 
   f.filename, 
@@ -91,8 +92,8 @@ export const getFileMetadataByProjectId = async (project_id: number) => {
   f.source,
   u.name AS uploader_name,
   u.surname AS uploader_surname 
-    FROM files f
-  JOIN users u ON f.uploaded_by = u.id
+    FROM ${tenant}.files f
+  JOIN public.users u ON f.uploaded_by = u.id
     WHERE project_id = :project_id 
     ORDER BY uploaded_time DESC, id ASC`;
   const result = await sequelize.query(query, {
